added can_manager, can_driver and can_worker. modified database_manager, instrument_repository

+++ database/repositories/instrument_repository.py
@@ -9,62 +9,153 @@ class InstrumentRepository:
     # ------------------------------------------------
     # ADD INSTRUMENT
     # ------------------------------------------------
+
     def add(self, instrument: InstrumentData):
 
         cursor = self.db.conn.cursor()
 
+        # Check if instrument already exists
         cursor.execute(
             """
-            INSERT INTO Instruments
-            (
-                InstrumentName,
-                Address,
-                InstrumentSNo,
-                CalibrationDueDate,
-                Status,
-                IsLocked
-            )
-            VALUES (?, ?, ?, ?, ?, ?)
+            SELECT InstrumentID
+            FROM Instruments
+            WHERE InstrumentSNo=?
             """,
-            (
-                instrument.instrument_name,
-                instrument.address,
-                instrument.instrument_sno,
-                instrument.calibration_due_date,
-                instrument.status,
-                instrument.is_locked
-            )
+            (instrument.instrument_sno,)
         )
 
-        instrument_id = cursor.lastrowid
+        row = cursor.fetchone()
 
-        # Channel mapping
-        if instrument.channel_id is not None:
+        if row:
+            instrument_id = row[0]
+        else:
             cursor.execute(
                 """
-                INSERT INTO ChannelInstrument
+                INSERT INTO Instruments
                 (
-                    ChannelID,
-                    InstrumentID,
-                    IsShared
+                    InstrumentName,
+                    Address,
+                    InstrumentSNo,
+                    CalibrationDueDate,
+                    Status,
+                    IsLocked
                 )
-                VALUES (?, ?, ?)
+                VALUES (?, ?, ?, ?, ?, ?)
                 """,
                 (
-                    instrument.channel_id,
-                    instrument_id,
-                    instrument.is_shared
+                    instrument.instrument_name,
+                    instrument.address,
+                    instrument.instrument_sno,
+                    instrument.calibration_due_date,
+                    instrument.status,
+                    instrument.is_locked
                 )
             )
 
+            instrument_id = cursor.lastrowid
+
+        # -----------------------------------
+        # Channel mapping
+        # -----------------------------------
+
+        channel_ids = [instrument.channel_id]
+
+        if instrument.is_shared:
+            if instrument.channel_id == 1:
+                channel_ids.append(2)
+            elif instrument.channel_id == 2:
+                channel_ids.append(1)
+
+        for channel_id in set(channel_ids):
+
+            cursor.execute(
+                """
+                SELECT 1
+                FROM ChannelInstrument
+                WHERE ChannelID=? AND InstrumentID=?
+                """,
+                (channel_id, instrument_id)
+            )
+
+            if cursor.fetchone() is None:
+                cursor.execute(
+                    """
+                    INSERT INTO ChannelInstrument
+                    (
+                        ChannelID,
+                        InstrumentID,
+                        IsShared
+                    )
+                    VALUES (?, ?, ?)
+                    """,
+                    (
+                        channel_id,
+                        instrument_id,
+                        instrument.is_shared
+                    )
+                )
+
         self.db.conn.commit()
-        print(f"Inserted Instrument ID: {instrument_id}")
 
         return instrument_id
+    # def add(self, instrument: InstrumentData):
+
+    #     cursor = self.db.conn.cursor()
+
+    #     cursor.execute(
+    #         """
+    #         INSERT INTO Instruments
+    #         (
+    #             InstrumentName,
+    #             Address,
+    #             InstrumentSNo,
+    #             CalibrationDueDate,
+    #             Status,
+    #             IsLocked
+    #         )
+    #         VALUES (?, ?, ?, ?, ?, ?)
+    #         """,
+    #         (
+    #             instrument.instrument_name,
+    #             instrument.address,
+    #             instrument.instrument_sno,
+    #             instrument.calibration_due_date,
+    #             instrument.status,
+    #             instrument.is_locked
+    #         )
+    #     )
+
+    #     instrument_id = cursor.lastrowid
+
+    #     # Channel mapping
+    #     if instrument.channel_id is not None:
+    #         cursor.execute(
+    #             """
+    #             INSERT INTO ChannelInstrument
+    #             (
+    #                 ChannelID,
+    #                 InstrumentID,
+    #                 IsShared
+    #             )
+    #             VALUES (?, ?, ?)
+    #             """,
+    #             (
+    #                 instrument.channel_id,
+    #                 instrument_id,
+    #                 instrument.is_shared
+    #             )
+    #         )
+
+    #     self.db.conn.commit()
+    #     print(f"Inserted Instrument ID: {instrument_id}")
+
+    #     return instrument_id
 
     # ------------------------------------------------
     # UPDATE INSTRUMENT
     # ------------------------------------------------
+
+
     def update(self, instrument: InstrumentData):
 
         cursor = self.db.conn.cursor()
@@ -92,7 +183,6 @@ class InstrumentRepository:
             )
         )
 
-        # Remove old mapping
         cursor.execute(
             """
             DELETE FROM ChannelInstrument
@@ -101,8 +191,15 @@ class InstrumentRepository:
             (instrument.instrument_id,)
         )
 
-        # Insert new mapping
-        if instrument.channel_id is not None:
+        channel_ids = [instrument.channel_id]
+
+        if instrument.is_shared:
+            if instrument.channel_id == 1:
+                channel_ids.append(2)
+            elif instrument.channel_id == 2:
+                channel_ids.append(1)
+
+        for channel_id in set(channel_ids):
             cursor.execute(
                 """
                 INSERT INTO ChannelInstrument
@@ -114,18 +211,102 @@ class InstrumentRepository:
                 VALUES (?, ?, ?)
                 """,
                 (
-                    instrument.channel_id,
+                    channel_id,
                     instrument.instrument_id,
                     instrument.is_shared
                 )
             )
 
         self.db.conn.commit()
+    # def update(self, instrument: InstrumentData):
+
+    #     cursor = self.db.conn.cursor()
+
+    #     cursor.execute(
+    #         """
+    #         UPDATE Instruments
+    #         SET
+    #             InstrumentName=?,
+    #             Address=?,
+    #             InstrumentSNo=?,
+    #             CalibrationDueDate=?,
+    #             Status=?,
+    #             IsLocked=?
+    #         WHERE InstrumentID=?
+    #         """,
+    #         (
+    #             instrument.instrument_name,
+    #             instrument.address,
+    #             instrument.instrument_sno,
+    #             instrument.calibration_due_date,
+    #             instrument.status,
+    #             instrument.is_locked,
+    #             instrument.instrument_id
+    #         )
+    #     )
+
+    #     # Remove old mapping
+    #     cursor.execute(
+    #         """
+    #         DELETE FROM ChannelInstrument
+    #         WHERE InstrumentID=?
+    #         """,
+    #         (instrument.instrument_id,)
+    #     )
+
+    #     # Insert new mapping
+    #     if instrument.channel_id is not None:
+    #         cursor.execute(
+    #             """
+    #             INSERT INTO ChannelInstrument
+    #             (
+    #                 ChannelID,
+    #                 InstrumentID,
+    #                 IsShared
+    #             )
+    #             VALUES (?, ?, ?)
+    #             """,
+    #             (
+    #                 instrument.channel_id,
+    #                 instrument.instrument_id,
+    #                 instrument.is_shared
+    #             )
+    #         )
+
+    #     self.db.conn.commit()
 
     # ------------------------------------------------
     # DELETE INSTRUMENT
     # ------------------------------------------------
-    def delete(self, instrument_id):
+    # def delete(self, instrument_id):
+
+    #     cursor = self.db.conn.cursor()
+
+    #     try:
+    #         cursor.execute(
+    #             """
+    #             DELETE FROM ChannelInstrument
+    #             WHERE InstrumentID=?
+    #             """,
+    #             (instrument_id,)
+    #         )
+
+    #         cursor.execute(
+    #             """
+    #             DELETE FROM Instruments
+    #             WHERE InstrumentID=?
+    #             """,
+    #             (instrument_id,)
+    #         )
+
+    #         self.db.conn.commit()
+
+    #     except Exception as e:
+    #         self.db.conn.rollback()
+    #         print("Delete Error:", e)
+    #         raise
+
+    def delete(self, instrument_id, channel_id):
 
         cursor = self.db.conn.cursor()
 
@@ -134,17 +315,30 @@ class InstrumentRepository:
                 """
                 DELETE FROM ChannelInstrument
                 WHERE InstrumentID=?
+                AND ChannelID=?
                 """,
-                (instrument_id,)
+                (instrument_id, channel_id)
             )
 
             cursor.execute(
                 """
-                DELETE FROM Instruments
+                SELECT COUNT(*)
+                FROM ChannelInstrument
                 WHERE InstrumentID=?
                 """,
                 (instrument_id,)
             )
+
+            count = cursor.fetchone()[0]
+
+            if count == 0:
+                cursor.execute(
+                    """
+                    DELETE FROM Instruments
+                    WHERE InstrumentID=?
+                    """,
+                    (instrument_id,)
+                )
 
             self.db.conn.commit()
